+++ CHANGELOG.md
@@ -8,10 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Changed (BREAKING)
+
 - Queries are no longer progressive by default. 
 - `ResultsProgressiveDisable()` has been removed.
 - Use `ResultsProgressiveEnabled()` to enable progressive queries.
 
+### Fixed
+
+- String quoting in default value of query parameters
 
 ## [0.14.2] - 2023-11-08
 
@@ -22,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.14.1] - 2023-09-27
 
 ### Added
+
 - Support new playfab domain
 
 ### Fixed